Added FileLoader and Player
Player holds the information and manipulators for the player object. FileLoader probably needs to change to just be a helper class/interface/whatever it's called as it's just a list of methods and doesn't need to be an initialized object. It handles loading data from XMLs into objects for other classes to use.